style/test suite: relocate test and fix typos
This moves tests which depend on mock Crust to the test folder.
