+++ Cargo.toml
@@ -10,7 +10,6 @@ readme = "README.md"
 repository = "https://github.com/maidsafe/safe_vault"
 version = "0.7.0"
 
-
 [dependencies]
 accumulator = "~0.3.0"
 chunk_store = "~0.4.0"
@@ -28,22 +27,21 @@ safe_network_common = "~0.1.1"
 sodiumoxide = "~0.0.10"
 xor_name = "~0.1.0"
 
-
 [build-dependencies]
 hyper = {version = "~0.9.1", optional = true}
 
-#[dev-dependencies.routing]
-#version = "*"
-#features = ["use-mock-crust"]
-
 [features]
-#default = ["use-mock-crust"]
 generate-diagrams = ["hyper"]
 use-mock-crust = ["routing/use-mock-crust"]
 
-
 [[test]]
 name = "mock_crust"
+
+[[test]]
+name = "data_manager"
+
+[[test]]
+name = "maid_manager"
 
 [[bin]]
 name = "safe_vault"
